pyproject.toml
[project]
name = "decision-intelligence-workflow"
version = "0.1.0"
description = "Decision intelligence workflow for AI and tech signal tracking"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4",
    "fastapi",
    "feedparser",
    "google-genai",
    "httpx",
    "pydantic-settings",
    "sqlalchemy",
    "uvicorn",
    "greenlet>=3.0.3,<3.1"
]
